+++ src/uwtools/config/core.py
@@ -2,547 +2,14 @@
 Core classes and functions for working with configs.
 """
 
-from __future__ import annotations
+from typing import List, Optional
 
-import configparser
-import copy
-import json
-import os
-import re
-from abc import ABC, abstractmethod
-from collections import OrderedDict, UserDict
-from io import StringIO
-from types import SimpleNamespace as ns
-from typing import Dict, List, Optional, Tuple, Type, Union
-
-import f90nml
-import yaml
-
-from uwtools.config.j2template import dereference
+from uwtools.config.formats.base import Config
+from uwtools.config.support import format_to_config, log_and_error
 from uwtools.exceptions import UWConfigError
 from uwtools.logging import MSGWIDTH, log
 from uwtools.types import DefinitePath, OptionalPath
-from uwtools.utils.file import FORMAT, get_file_type, readable, writable
-
-_INCLUDE_TAG = "!INCLUDE"
-
-_MSGS = ns(
-    unhashable="""
-ERROR:
-The input config file may contain a Jinja2 templated value at the location listed above. Ensure the
-value is enclosed in quotes.
-""".strip(),
-    unregistered_constructor="""
-ERROR:
-The input config file contains a constructor that is not registered with the uwtools package.
-
-constructor: {constructor}
-config file: {config_file}
-
-Define the constructor before proceeding.
-""".strip(),
-    unregistered_filter="""
-ERROR:
-The input config file contains a Jinja2 filter that is not registered with the uwtools package.
-
-filter: {filter}
-key: {key}
-
-Define the filter before proceeding.
-""".strip(),
-)
-
-
-class Config(ABC, UserDict):
-    """
-    A base class specifying (and partially implementing) methods to read, manipulate, and write
-    several configuration-file formats.
-    """
-
-    def __init__(self, config_file: OptionalPath = None) -> None:
-        """
-        Construct a Config object.
-
-        :param config_file: Path to the config file to load.
-        """
-        super().__init__()
-        self._config_file = str(config_file) if config_file else None
-        self.update(self._load(self._config_file))
-
-    def __repr__(self) -> str:
-        """
-        The string representation of a Config object.
-        """
-        return json.dumps(self.data)
-
-    # Private methods
-
-    @abstractmethod
-    def _load(self, config_file: OptionalPath) -> dict:
-        """
-        Reads and parses a config file.
-
-        Returns the result of loading and parsing the specified config file, or stdin if no file is
-        given.
-
-        :param config_file: Path to config file to load.
-        """
-
-    def _load_paths(self, config_files: List[str]) -> dict:
-        """
-        Merge and return the contents of a collection of config files.
-
-        :param config_files: Paths to the config files to read and merge.
-        """
-        cfg = {}
-        for config_file in config_files:
-            if not os.path.isabs(config_file):
-                if self._config_file:
-                    config_file = os.path.join(os.path.dirname(self._config_file), config_file)
-                else:
-                    raise _log_and_error(
-                        "Reading from stdin, a relative path was encountered: %s" % config_file
-                    )
-            cfg.update(self._load(config_file=config_file))
-        return cfg
-
-    # Public methods
-
-    def characterize_values(self, values: dict, parent: str) -> Tuple[list, list, list]:
-        """
-        Characterize values as complete, empty, or template placeholders.
-
-        :param values: The dictionary to examine.
-        :param parent: Parent key.
-        """
-        complete: List[str] = []
-        empty: List[str] = []
-        template: List[str] = []
-        for key, val in values.items():
-            if isinstance(val, dict):
-                complete.append(f"    {parent}{key}")
-                c, e, t = self.characterize_values(val, f"{parent}{key}.")
-                complete, empty, template = complete + c, empty + e, template + t
-            elif isinstance(val, list):
-                for item in val:
-                    if isinstance(item, dict):
-                        c, e, t = self.characterize_values(item, parent)
-                        complete, empty, template = complete + c, empty + e, template + t
-                        complete.append(f"    {parent}{key}")
-                    elif "{{" in str(val) or "{%" in str(val):
-                        template.append(f"    {parent}{key}: {val}")
-                        break
-            elif "{{" in str(val) or "{%" in str(val):
-                template.append(f"    {parent}{key}: {val}")
-            elif val == "" or val is None:
-                empty.append(f"    {parent}{key}")
-            else:
-                complete.append(f"    {parent}{key}")
-        return complete, empty, template
-
-    def compare_config(self, dict1: dict, dict2: Optional[dict] = None) -> bool:
-        """
-        Compare two config dictionaries.
-
-        Assumes a section/key/value structure.
-
-        :param dict1: The first dictionary.
-        :param dict2: The second dictionary.
-        """
-        dict2 = self.data if dict2 is None else dict2
-        diffs: dict = {}
-
-        for sect, items in dict2.items():
-            for key, val in items.items():
-                if val != dict1.get(sect, {}).get(key, ""):
-                    try:
-                        diffs[sect][key] = f" - {val} + {dict1.get(sect, {}).get(key)}"
-                    except KeyError:
-                        diffs[sect] = {}
-                        diffs[sect][key] = f" - {val} + {dict1.get(sect, {}).get(key)}"
-
-        for sect, items in dict1.items():
-            for key, val in items.items():
-                if val != dict2.get(sect, {}).get(key, "") and diffs.get(sect, {}).get(key) is None:
-                    try:
-                        diffs[sect][key] = f" - {dict2.get(sect, {}).get(key)} + {val}"
-                    except KeyError:
-                        diffs[sect] = {}
-                        diffs[sect][key] = f" - {dict2.get(sect, {}).get(key)} + {val}"
-
-        for sect, keys in diffs.items():
-            for key in keys:
-                msg = f"{sect}: {key:>15}: {keys[key]}"
-                log.info(msg)
-
-        return not diffs
-
-    @property
-    def depth(self) -> int:
-        """
-        The depth of this config's hierarchy.
-        """
-        return depth(self.data)
-
-    def dereference_all(self) -> None:
-        """
-        Render as much Jinja2 syntax as possible.
-        """
-        while True:
-            new = dereference(val=self.data, context={**os.environ, **self.data})
-            if new == self.data:
-                break
-            assert isinstance(new, dict)
-            self.data = new
-
-    @abstractmethod
-    def dump(self, path: OptionalPath) -> None:
-        """
-        Dumps the config to stdout or a file.
-
-        :param path: Path to dump config to.
-        """
-
-    @staticmethod
-    @abstractmethod
-    def dump_dict(path: OptionalPath, cfg: dict, opts: Optional[ns] = None) -> None:
-        """
-        Dumps a provided config dictionary to stdout or a file.
-
-        :param path: Path to dump config to.
-        :param cfg: The in-memory config object to dump.
-        :param opts: Other options required by a subclass.
-        """
-
-    def parse_include(self, ref_dict: Optional[dict] = None) -> None:
-        """
-        Recursively process include directives in a config object.
-
-        Recursively traverse the dictionary, replacing include tags with the contents of the files
-        they specify. Assumes a section/key/value structure. YAML provides this functionality in its
-        own loader.
-
-        :param ref_dict: A config object to process instead of the object's own data.
-        """
-        if ref_dict is None:
-            ref_dict = self.data
-        for key, value in copy.deepcopy(ref_dict).items():
-            if isinstance(value, dict):
-                self.parse_include(ref_dict[key])
-            elif isinstance(value, str):
-                if m := re.match(r"^\s*%s\s+(.*)" % _INCLUDE_TAG, value):
-                    filepaths = yaml.safe_load(m[1])
-                    self.update_values(self._load_paths(filepaths))
-                    del ref_dict[key]
-
-    def update_values(self, src: Union[dict, Config], dst: Optional[Config] = None):
-        """
-        Updates a Config object.
-
-        Update the instance's own data (or, optionally, that of the specifed Config object) with the
-        values provided by another dictionary or Config object.
-
-        :param src: The dictionary with new data to use.
-        :param dst: The Config to update with the new data.
-        """
-        srcdict = src.data if isinstance(src, Config) else src
-        dstcfg = self if dst is None else dst
-        for key, new_val in srcdict.items():
-            if isinstance(new_val, dict):
-                if isinstance(dstcfg.get(key), dict):
-                    self.update_values(new_val, dstcfg[key])
-                else:
-                    dstcfg[key] = new_val
-            else:
-                dstcfg[key] = new_val
-
-
-class INIConfig(Config):
-    """
-    Concrete class to handle INI config files.
-    """
-
-    def __init__(
-        self,
-        config_file: str,
-        space_around_delimiters: bool = True,
-    ):
-        """
-        Construct an INIConfig object.
-
-        Spaces may be included for INI format, but should be excluded for bash.
-
-        :param config_file: Path to the config file to load.
-        :param space_around_delimiters: Include spaces around delimiters?
-        """
-        super().__init__(config_file)
-        self.space_around_delimiters = space_around_delimiters
-        self.parse_include()
-
-    # Private methods
-
-    def _load(self, config_file: OptionalPath) -> dict:
-        """
-        Reads and parses an INI file.
-
-        See docs for Config._load().
-
-        :param config_file: Path to config file to load.
-        """
-        # The protected _sections method is the most straightforward way to get at the dict
-        # representation of the parse config.
-
-        cfg = configparser.ConfigParser()
-        cfg.optionxform = str  # type: ignore
-        sections = cfg._sections  # type: ignore # pylint: disable=protected-access
-        with readable(config_file) as f:
-            raw = f.read()
-        try:
-            cfg.read_string(raw)
-            return dict(sections)
-        except configparser.MissingSectionHeaderError:
-            cfg.read_string("[top]\n" + raw)
-            return dict(sections.get("top"))
-
-    # Public methods
-
-    def dump(self, path: OptionalPath) -> None:
-        """
-        Dumps the config in INI format.
-
-        :param path: Path to dump config to.
-        """
-        INIConfig.dump_dict(path, self.data, ns(space=self.space_around_delimiters))
-
-    @staticmethod
-    def dump_dict(path: OptionalPath, cfg: dict, opts: Optional[ns] = None) -> None:
-        """
-        Dumps a provided config dictionary in INI or sh format.
-
-        :param path: Path to dump config to.
-        :param cfg: The in-memory config object to dump.
-        :param space_around_delimiters: Place spaces around delimiters?
-        """
-        # Configparser adds a newline after each section, presumably to create nice-looking output
-        # when an INI contains multiple sections. Unfortunately, it also adds a newline after the
-        # _final_ section, resulting in an anomalous trailing newline. To avoid this, write first to
-        # memory, then strip the trailing newline.
-        parser = configparser.ConfigParser()
-        s = StringIO()
-        cfgdepth = depth(cfg)
-        assert cfgdepth in (1, 2)  # 1 => .sh, 2 => .ini
-        if cfgdepth == 2:
-            parser.read_dict(cfg)
-            parser.write(s, space_around_delimiters=opts.space if opts else True)
-        else:
-            space = " " if not opts or opts.space else ""
-            for key, value in cfg.items():
-                print(f"{key}{space}={space}{value}", file=s)
-        with writable(path) as f:
-            print(s.getvalue().strip(), file=f)
-        s.close()
-
-
-class NMLConfig(Config):
-    """
-    Concrete class to handle Fortran namelist files.
-    """
-
-    def __init__(self, config_file) -> None:
-        super().__init__(config_file)
-        self.parse_include()
-
-    # Private methods
-
-    def _load(self, config_file: OptionalPath) -> dict:
-        """
-        Reads and parses a Fortran namelist file.
-
-        See docs for Config._load().
-
-        :param config_file: Path to config file to load.
-        """
-
-        # f90nml returns OrderedDict objects to maintain the order of namelists in the namelist
-        # files that it reads. But in Python 3.6+ the standard dict maintains order as well. Since
-        # OrderedDict can cause problems downstream when serializing to YAML, convert OrderedDict
-        # objects to standard dicts here.
-
-        def from_od(d):
-            return {key: from_od(val) if isinstance(val, dict) else val for key, val in d.items()}
-
-        with readable(config_file) as f:
-            return from_od(f90nml.read(f).todict())
-
-    # Public methods
-
-    def dump(self, path: OptionalPath) -> None:
-        """
-        Dumps the config in Fortran namelist format.
-
-        :param path: Path to dump config to.
-        """
-        NMLConfig.dump_dict(path, self.data)
-
-    @staticmethod
-    def dump_dict(path: OptionalPath, cfg: dict, opts: Optional[ns] = None) -> None:
-        """
-        Dumps a provided config dictionary in Fortran namelist format.
-
-        :param path: Path to dump config to.
-        :param cfg: The in-memory config object to dump.
-        :param opts: Other options required by a subclass.
-        """
-
-        # f90nml honors namelist and variable order if it receives an OrderedDict as input, so
-        # ensure that it receives one.
-
-        def to_od(d):
-            return OrderedDict(
-                {key: to_od(val) if isinstance(val, dict) else val for key, val in d.items()}
-            )
-
-        with writable(path) as f:
-            f90nml.Namelist(to_od(cfg)).write(f, sort=False)
-
-
-class YAMLConfig(Config):
-    """
-    Concrete class to handle YAML config files.
-    """
-
-    def __repr__(self) -> str:
-        """
-        The string representation of a YAMLConfig object.
-        """
-        return yaml.dump(self.data)
-
-    # Private methods
-
-    def _load(self, config_file: OptionalPath) -> dict:
-        """
-        Reads and parses a YAML file.
-
-        See docs for Config._load().
-
-        :param config_file: Path to config file to load.
-        """
-        loader = self._yaml_loader
-        with readable(config_file) as f:
-            try:
-                return yaml.load(f.read(), Loader=loader)
-            except yaml.constructor.ConstructorError as e:
-                if e.problem:
-                    if "unhashable" in e.problem:
-                        msg = _MSGS.unhashable
-                    else:
-                        constructor = e.problem.split()[-1]
-                        msg = _MSGS.unregistered_constructor.format(
-                            config_file=config_file, constructor=constructor
-                        )
-                else:
-                    msg = str(e)
-                raise _log_and_error(msg) from e
-
-    def _yaml_include(self, loader: yaml.Loader, node: yaml.SequenceNode) -> dict:
-        """
-        Returns a dictionary with include tags processed.
-
-        :param loader: The YAML loader.
-        :param node: A YAML node.
-        """
-        filepaths = loader.construct_sequence(node)
-        return self._load_paths(filepaths)
-
-    @property
-    def _yaml_loader(self) -> type[yaml.SafeLoader]:
-        """
-        Set up the loader with the appropriate constructors.
-        """
-        loader = yaml.SafeLoader
-        loader.add_constructor(_INCLUDE_TAG, self._yaml_include)
-        return loader
-
-    # Public methods
-
-    def dump(self, path: OptionalPath) -> None:
-        """
-        Dumps the config in YAML format.
-
-        :param path: Path to dump config to.
-        """
-        YAMLConfig.dump_dict(path, self.data)
-
-    @staticmethod
-    def dump_dict(path: OptionalPath, cfg: dict, opts: Optional[ns] = None) -> None:
-        """
-        Dumps a provided config dictionary in YAML format.
-
-        :param path: Path to dump config to.
-        :param cfg: The in-memory config object to dump.
-        :param opts: Other options required by a subclass.
-        """
-        with writable(path) as f:
-            yaml.dump(cfg, f, sort_keys=False)
-
-
-class FieldTableConfig(YAMLConfig):
-    """
-    This class exists to write out a field_table format given that its configuration has been set by
-    an input YAML file.
-    """
-
-    # Public methods
-
-    def dump(self, path: OptionalPath) -> None:
-        """
-        Dumps the config in Field Table format.
-
-        :param path: Path to dump config to.
-        """
-        FieldTableConfig.dump_dict(path, self.data)
-
-    @staticmethod
-    def dump_dict(path: OptionalPath, cfg: dict, opts: Optional[ns] = None) -> None:
-        """
-        Dumps a provided config dictionary in Field Table format.
-
-        FMS field and tracer managers must be registered in an ASCII table called 'field_table'.
-        This table lists field type, target model and methods the querying model will ask for. See
-        UFS documentation for more information:
-
-        https://ufs-weather-model.readthedocs.io/en/ufs-v1.0.0/InputsOutputs.html#field-table-file
-
-        The example format for generating a field file is:
-
-        sphum:
-          longname: specific humidity
-          units: kg/kg
-          profile_type:
-            name: fixed
-            surface_value: 1.e30
-
-        :param path: Path to dump config to.
-        :param cfg: The in-memory config object to dump.
-        :param opts: Other options required by a subclass.
-        """
-        lines = []
-        for field, settings in cfg.items():
-            lines.append(f' "TRACER", "atmos_mod", "{field}"')
-            for key, value in settings.items():
-                if isinstance(value, dict):
-                    method_string = f'{" ":7}"{key}", "{value.pop("name")}"'
-                    # All control vars go into one set of quotes.
-                    control_vars = [f"{method}={val}" for method, val in value.items()]
-                    # Whitespace after the comma matters.
-                    lines.append(f'{method_string}, "{", ".join(control_vars)}"')
-                else:
-                    # Formatting of variable spacing dependent on key length.
-                    lines.append(f'{" ":11}"{key}", "{value}"')
-            lines[-1] += " /"
-        with writable(path) as f:
-            print("\n".join(lines), file=f)
-
+from uwtools.utils.file import FORMAT, get_file_type
 
 # Public functions
 
@@ -571,37 +38,6 @@ def compare_configs(
     return cfg_a.compare_config(cfg_b.data)
 
 
-def depth(d: dict) -> int:
-    """
-    The depth of a dictionary.
-
-    :param d: The dictionary whose depth to calculate.
-    :return: The length of the longest path to a value in the dictionary.
-    """
-    return (max(map(depth, d.values())) + 1) if isinstance(d, dict) else 0
-
-
-def format_to_config(cli_name: str) -> Type[Config]:
-    """
-    Maps a CLI format name to its corresponding Config class.
-
-    :param cli_name: The format name as known to the CLI.
-    :return: The appropriate Config class.
-    """
-    classes: Dict[str, type[Config]] = {
-        FORMAT.fieldtable: FieldTableConfig,
-        FORMAT.ini: INIConfig,
-        FORMAT.nml: NMLConfig,
-        FORMAT.yaml: YAMLConfig,
-    }
-    try:
-        return classes[cli_name]
-    except KeyError as e:
-        raise _log_and_error(
-            "Format '%s' should be one of: %s" % (e.args[0], ", ".join(classes.keys()))
-        ) from e
-
-
 def print_config_section(config: dict, key_path: List[str]) -> None:
     """
     Descends into the config via the given keys, then prints the contents of the located subtree as
@@ -615,14 +51,14 @@ def print_config_section(config: dict, key_path: List[str]) -> None:
         try:
             subconfig = config[section]
         except KeyError as e:
-            raise _log_and_error(f"Bad config path: {current_path}") from e
+            raise log_and_error(f"Bad config path: {current_path}") from e
         if not isinstance(subconfig, dict):
-            raise _log_and_error(f"Value at {current_path} must be a dictionary")
+            raise log_and_error(f"Value at {current_path} must be a dictionary")
         config = subconfig
     output_lines = []
     for key, value in config.items():
         if type(value) not in (bool, float, int, str):
-            raise _log_and_error(f"Non-scalar value {value} found at {current_path}")
+            raise log_and_error(f"Non-scalar value {value} found at {current_path}")
         output_lines.append(f"{key}={value}")
     print("\n".join(sorted(output_lines)))
 
@@ -666,16 +102,6 @@ def realize_config(
 
 
 # Private functions
-
-
-def _log_and_error(msg: str) -> Exception:
-    """
-    Log an error message and raise an exception.
-
-    :param msg: The error message to log and to associate with raised exception.
-    """
-    log.error(msg)
-    return UWConfigError(msg)
 
 
 def _realize_config_check_depths(input_obj: Config, output_format: str) -> None:
