+++ src/uwtools/config/support.py
@@ -1,0 +1,46 @@
+from importlib import import_module
+from typing import Type
+
+from uwtools.exceptions import UWConfigError
+from uwtools.logging import log
+from uwtools.utils.file import FORMAT
+
+INCLUDE_TAG = "!INCLUDE"
+
+
+def depth(d: dict) -> int:
+    """
+    The depth of a dictionary.
+
+    :param d: The dictionary whose depth to calculate.
+    :return: The length of the longest path to a value in the dictionary.
+    """
+    return (max(map(depth, d.values())) + 1) if isinstance(d, dict) else 0
+
+
+def format_to_config(fmt: str) -> Type:
+    """
+    Maps a CLI format name to its corresponding Config class.
+
+    :param fmt: The format name as known to the CLI.
+    :return: The appropriate Config class.
+    """
+    lookup = {
+        FORMAT.fieldtable: "FieldTableConfig",
+        FORMAT.ini: "INIConfig",
+        FORMAT.nml: "NMLConfig",
+        FORMAT.yaml: "YAMLConfig",
+    }
+    if not fmt in lookup:
+        raise log_and_error("Format '%s' should be one of: %s" % (fmt, ", ".join(lookup)))
+    return getattr(import_module(f"uwtools.config.formats.{fmt}"), lookup[fmt])
+
+
+def log_and_error(msg: str) -> Exception:
+    """
+    Log an error message and raise an exception.
+
+    :param msg: The error message to log and to associate with raised exception.
+    """
+    log.error(msg)
+    return UWConfigError(msg)
